modify project for helping setup card reader
