+++ CardReaderTestApp/App.config
