+++ CardReaderTestApp/CardReaderTestApp.csproj
@@ -6,8 +6,8 @@
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
     <ProjectGuid>{DE879C39-E74F-4D75-B53D-0E1F263DEBCF}</ProjectGuid>
     <OutputType>WinExe</OutputType>
-    <RootNamespace>thai_id_card_reader_window_app</RootNamespace>
-    <AssemblyName>thai-id-card-reader-window-app</AssemblyName>
+    <RootNamespace>CardReaderTestApp</RootNamespace>
+    <AssemblyName>CardReaderTestApp</AssemblyName>
     <TargetFrameworkVersion>v4.7.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
@@ -80,6 +80,7 @@
     <Compile Include="Properties\Resources.Designer.cs">
       <AutoGen>True</AutoGen>
       <DependentUpon>Resources.resx</DependentUpon>
+      <DesignTime>True</DesignTime>
     </Compile>
     <None Include="manifest.json" />
     <None Include="Properties\Settings.settings">
