+++ CardReaderTestApp/Models/ResponseModel.cs
