+++ CardReaderTestApp/Program.cs
