+++ CardReaderTestApp/Properties/Resources.resx
