+++ CardReaderTestApp/Properties/Settings.settings
