+++ CardReaderTestApp/frmMain.resx
