+++ CardReaderTestApp/manifest.json
