+++ thai-id-card-reader.sln
@@ -1,11 +1,11 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31702.278
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.32014.148
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "thai-id-card-reader-window-app", "thai-id-card-reader-window-app\thai-id-card-reader-window-app.csproj", "{DE879C39-E74F-4D75-B53D-0E1F263DEBCF}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "thai-id-card-reader-console-app", "thai-id-card-reader-console-app\thai-id-card-reader-console-app.csproj", "{A7CF0C6E-68EE-449E-82CF-5482C5FE20FB}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CardReaderTestApp", "CardReaderTestApp\CardReaderTestApp.csproj", "{DE879C39-E74F-4D75-B53D-0E1F263DEBCF}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -13,14 +13,14 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{DE879C39-E74F-4D75-B53D-0E1F263DEBCF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{DE879C39-E74F-4D75-B53D-0E1F263DEBCF}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{DE879C39-E74F-4D75-B53D-0E1F263DEBCF}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{DE879C39-E74F-4D75-B53D-0E1F263DEBCF}.Release|Any CPU.Build.0 = Release|Any CPU
 		{A7CF0C6E-68EE-449E-82CF-5482C5FE20FB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{A7CF0C6E-68EE-449E-82CF-5482C5FE20FB}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A7CF0C6E-68EE-449E-82CF-5482C5FE20FB}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A7CF0C6E-68EE-449E-82CF-5482C5FE20FB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DE879C39-E74F-4D75-B53D-0E1F263DEBCF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DE879C39-E74F-4D75-B53D-0E1F263DEBCF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DE879C39-E74F-4D75-B53D-0E1F263DEBCF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DE879C39-E74F-4D75-B53D-0E1F263DEBCF}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
